logging_service/requirements.txt
Django>=5.0
djangorestframework
mysqlclient
nats-py
requests
django-cors-headers
